Histogramm-Routinen hinzugefügt, Verbesserung in clean_rundir.sh

+++ Scripts/clean_rundir.sh
@@ -3,4 +3,9 @@
 WORKINGDIR=$(pwd)
 RUNDIR=$WORKINGDIR/$1
 
+if [[ $1 == "" ]]; then
+    echo "usage: ./clean_rundir.sh <directory>"
+    exit -1
+fi
+
 find $RUNDIR ! \( -name '*.slha' -o -name '*.input' -o -name 'pwgseeds.dat' \) -type f -exec rm -f {} +
